requirements.txt
python-docx
langchain
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
docx2txt
psycopg2-binary